Clean up ModServiceCard, rename to LabelingServiceCard
